requirements.txt
# ════════════════════════════════════════════════
# GAMMA PLATFORM — requirements.txt
# Python 3.11+  |  pip install -r requirements.txt
# ════════════════════════════════════════════════

# ── Core API Framework ───────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==13.1
pydantic==2.9.0
pydantic-settings==2.5.0

# ── Database ─────────────────────────────────────
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# ── Voice Pipeline: STT ──────────────────────────
deepgram-sdk==3.7.0

# ── Voice Pipeline: LLM ──────────────────────────
google-generativeai==0.8.3

# ── Voice Pipeline: TTS (Google — add when GCP ready) ────
google-cloud-texttospeech==2.17.2
google-auth==2.35.0

# ── Telephony ────────────────────────────────────
twilio==9.3.4

# ── File Processing (Knowledge Base) ─────────────
pypdf==5.0.1
python-docx==1.1.2
aiofiles==24.1.0
python-magic==0.4.27

# ── Auth & Security ──────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# ── HTTP ─────────────────────────────────────────
httpx==0.27.2
authlib==1.3.2

# ── Task Queue & Scheduling ──────────────────────
celery==5.4.0
redis==5.1.1
apscheduler==3.10.4
kombu==5.3.7

# ── CSV / Data Processing ────────────────────────
pandas==2.2.3
chardet==5.2.0

# ── Monitoring & Logging ─────────────────────────
psutil==6.1.0
structlog==24.4.0
prometheus-client==0.21.0

# ── Environment ──────────────────────────────────
python-dotenv==1.0.1

# ── Testing ──────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
factory-boy==3.3.1
